requirements.txt
numpy==1.8.2
scipy==0.19.1
tensorflow-gpu==1.13.1
keras==2.1.4
